Fixed bug with log level I/O
This commit include the change inside the default value if log level from IO to debug.

In addition there is a limitation inside the lines print to edit text because if I push too text inside the Edit text, it will freez the UI thread. This cases happened too with log level to IO

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>