Fixed crash described in #54
This PR contains the code to fixes the issue #54.

The crash happened because during the init node state is not possible to change the activity because the change activity action stops the previous activity, such as MainActivity to run the LogActiviry or SettingActivity.

So, the solution is to make a more solid check of the node running and disable the setting and log menu item during the startup of the node.

In addition, I added propriety to MainActivity such as singleTop, this propriety force to use the instance of activity created, if exist. This resolved the UI bug when using the button in App bar to come back to MainActivity

In conclusion, I added also a little of clean up code!

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>