Introduce ViewModel and Courutines
This commit implement contains the update of all dependencies inside the app.

In addition, it includes Kotlin Courutines instead of org.jetbrains.anko library that looks like archived.

In conclusion, it includes some fixed of the first review, such as

- if the log is huge, it takes some time to read at chunks with 25ms delay for each
 - after fetching logs, tor looks to append a newline at the end, but not the lightning log.
 - remote some verbosity "ADDED new line to log view"
 - Refactoring code inside the class TorServices.kt as mentioned inside #85

But there is some work to do to fix the issue reported below. At the moment this behavior is avoided with a simple trick to start from the top to scrolling but is very bad when the log is huge

- if I scroll up to see most older logs, I see the recent one and never reach the starting point.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>