Add kotlin linter plugin to and update all the code
This code introduces a koltin linter, in particular a plugin called  kotlinter, and it is available on  https://github.com/jeremymailen/kotlinter-gradle

This commit adds the support in all the app and also in the Github cli, so now all the commit need to pass also the kotlin format check.

To make easy the support of the guideline a .editorconfig is added but an additional configuration of Android studio is needed.

The Android Studio configuration that is needed is 

Settings -> Editor -> CodeStyle -> Kotlin -> import -> enable  Use single name import.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>